Anti-Cheat
Added first anti-cheat check. While this system is very primitive and only checks for a maximum speed, it's a first step in the right direction. The implementation enables easy addition to further checks. Depending on how complex the system becomes, it might be necessary to add a history for every player's position over a longer period of time. This could be useful for checking behaviour that is just slightly above normal, but not blatantly obvious from one packet to the next.

Note:
client.py is only changed because I had to use a new server address.